requirement.txt
asgiref==3.3.1
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
defusedxml==0.7.1
Django==3.1.5
django-crispy-forms==1.11.1
django-filter==21.1
djangorestframework==3.13.0
httpie==2.6.0
idna==3.3
importlib-metadata==4.10.0
Markdown==3.3.6
Pillow==8.1.0
Pygments==2.10.0
PySocks==1.7.1
pytz==2020.5
PyYAML==6.0
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.1
urllib3==1.26.7
zipp==3.6.0
